backend/requirements.txt
# FastAPI 核心框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 数据库相关
sqlalchemy==2.0.23
alembic==1.12.1
sqlite3  # 内置，无需安装

# 认证和安全
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# 数据验证和序列化
pydantic==2.4.2
pydantic-settings==2.0.3
email-validator==2.1.0

# HTTP客户端
httpx==0.25.2
aiohttp==3.9.1

# WebSocket支持
websockets==12.0

# 日志和监控
python-json-logger==2.0.7
structlog==23.2.0

# 工具库
python-dotenv==1.0.0
pytz==2023.3
croniter==2.0.1

# 开发和测试工具
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.10.1
flake8==6.1.0
mypy==1.7.0

# CORS支持
python-cors==1.7.0

# 缓存（可选）
redis==5.0.1
aioredis==2.0.1

# 文件处理
pillow==10.1.0
python-magic==0.4.27

# 配置管理
pyyaml==6.0.1
configparser==6.0.0
